\name{supported_formats}
\alias{supported_formats}
\title{List Supported and Unsupported File Formats}
\description{
  Displays a list of supported and unsupported file formats for the \code{auto_read()} and \code{auto_write()} functions.
}
\usage{
supported_formats()
}
\value{
  Prints a list of file formats supported and unsupported by the autofileIO package with tick indicating supported formats and cross indicating unsupported formats. It does not return a value.
}
\examples{
supported_formats()
}